+++ packages/framework/src/Model/Product/Search/SearchClient.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Shopsys\FrameworkBundle\Model\Product\Search;
+
+interface SearchClient
+{
+    /**
+     * @param int $domainId
+     * @param string|null $searchText
+     * @return int[]
+     */
+    public function search(int $domainId, $searchText): array;
+}